requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli==1.32.111
flake8
bandit
python-dotenv>=0.5.1

# DL
ultralytics
albumentations
tqdm
